requirements.txt
SciPy 
TensorFlow==1.15.0
scikit-learn==0.21.3  
matplotlib==3.1.1 
NLtk==3.5 
h5py==2.10 
pandas 
numpy==1.19.5
pysster
pygad
progressbar2 
torchvision
viennarna==2.5.1